# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Финни — API справочников и прогресса",
    "description": "\nБэкенд игры «Питомец Финни».\n\n**Приложение работает без сервера.** Игра полностью офлайн: контент зашит в\nкод приложения, прогресс хранится на устройстве. Этот сервис необязательный —\nон нужен для двух вещей:\n\n1. **Справочники** (`/api/v1/content/*`) — тот же контент, что в приложении:\n   магазин, цели копилки, задания дорог «Математика» и «Финансы», питомцы и\n   правила экономики. JSON собирается из Dart-кода скриптом\n   `scripts/sync_from_mobile.py`. Разделы кэшируются по ETag\n   (`If-None-Match` → 304).\n2. **Прогресс** (`/api/v1/progress/*`) — резервная копия снимка локального\n   профиля (переустановка, смена устройства, проверка экспертом).\n\nПерсональные данные не собираются: профиль адресуется UUID, который\nгенерирует само приложение.\n",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/progress/{profile_id}": {
      "put": {
        "tags": [
          "Прогресс"
        ],
        "summary": "Выгрузить прогресс на сервер",
        "description": "Клиент присылает снимок целиком и увеличивает `revision` при каждой выгрузке. Если на сервере лежит бóльшая ревизия — 409 вместе с серверным снимком, чтобы приложение могло сравнить их и не потерять прогресс. Перезаписать принудительно — `?force=true`.",
        "operationId": "upload_progress_api_v1_progress__profile_id__put",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID локального профиля, который генерирует приложение",
              "title": "Profile Id"
            },
            "description": "UUID локального профиля, который генерирует приложение"
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Перезаписать серверную версию",
              "default": false,
              "title": "Force"
            },
            "description": "Перезаписать серверную версию"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressSnapshot"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressAck"
                }
              }
            }
          },
          "201": {
            "description": "Профиль сохранён впервые",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressAck"
                }
              }
            }
          },
          "409": {
            "description": "На сервере более новая ревизия",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressConflict"
                }
              }
            }
          },
          "422": {
            "description": "Снимок нарушает правила ТЗ (например, отрицательный баланс)"
          }
        }
      },
      "get": {
        "tags": [
          "Прогресс"
        ],
        "summary": "Забрать прогресс с сервера",
        "description": "Нужен при переустановке приложения или смене устройства.",
        "operationId": "download_progress_api_v1_progress__profile_id__get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID локального профиля, который генерирует приложение",
              "title": "Profile Id"
            },
            "description": "UUID локального профиля, который генерирует приложение"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredProgress"
                }
              }
            }
          },
          "404": {
            "description": "Прогресс для этого профиля не выгружался"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Прогресс"
        ],
        "summary": "Удалить серверную копию прогресса",
        "description": "Для раздела взрослого (ТЗ 3.5: сброс и удаление данных доступны без обращения к разработчику). Операция идемпотентна: если копии нет, ответ всё равно 204.",
        "operationId": "delete_progress_api_v1_progress__profile_id__delete",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID локального профиля, который генерирует приложение",
              "title": "Profile Id"
            },
            "description": "UUID локального профиля, который генерирует приложение"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "ref"
        ],
        "title": "ContentWarning",
        "description": "Мягкое предупреждение о ссылке на неизвестный серверу контент.\n\nНе ошибка: у приложения может быть более новый контент, и терять из-за\nэтого прогресс нельзя."
      },
...
      "GoalState": {
        "properties": {
          "goal_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goal Id"
          },
          "title": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Title"
          },
          "emoji": {
            "type": "string",
            "minLength": 1,
            "title": "Emoji"
          },
          "target": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 50.0,
            "title": "Target"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "emoji",
          "target"
        ],
        "title": "GoalState",
        "description": "Цель копилки: готовая (`goal_id`) или своя (`goal_id: null`)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InventoryEntry": {
        "properties": {
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "quantity": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Quantity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "item_id",
          "quantity"
        ],
        "title": "InventoryEntry"
      },
...
      "LessonsState": {
        "properties": {
          "solved": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Solved"
          },
          "retry": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Retry",
            "description": "Задачи «Повтори»"
          },
          "mistakes": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Mistakes",
            "default": 0
          },
          "last_solved_day": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Last Solved Day",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "LessonsState",
        "description": "Задания дорог. Ошибка не отнимает прогресс — задание уходит в «Повтори»."
      },
...
      "PetState": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 15,
            "minLength": 1,
            "title": "Name"
          },
          "species_id": {
            "type": "string",
            "title": "Species Id",
            "description": "cat | dog | penguin"
          },
          "variant_id": {
            "type": "string",
            "title": "Variant Id",
            "description": "Окраска: v1 | v2 | v3"
          },
          "level": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Level"
          },
          "xp": {
            "type": "integer",
            "maximum": 99.0,
            "minimum": 0.0,
            "title": "Xp",
            "description": "Опыт внутри текущего уровня"
          },
          "stats": {
            "$ref": "#/components/schemas/PetStatsState"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "species_id",
          "variant_id",
          "level",
          "xp",
          "stats"
        ],
        "title": "PetState"
      },
      "PetStatsState": {
        "properties": {
          "satiety": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Satiety",
            "description": "В приложении — hunger"
          },
          "happiness": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Happiness"
          },
          "cleanliness": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Cleanliness"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "satiety",
          "happiness",
          "cleanliness"
        ],
        "title": "PetStatsState"
      },
...
      "PlayerState": {
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 15,
            "minLength": 1,
            "title": "Nickname"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 10.0,
                "minimum": 7.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age",
            "description": "7, 8, 9 или 10 («10+»); null — старый сейв"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nickname"
        ],
        "title": "PlayerState"
      },
      "ProfileSettings": {
        "properties": {
          "dark_theme": {
            "type": "boolean",
            "title": "Dark Theme",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProfileSettings"
      },
      "ProgressAck": {
        "properties": {
          "profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Profile Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "server_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Server Updated At"
          },
          "created": {
            "type": "boolean",
            "title": "Created",
            "description": "true — профиль сохранён впервые"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ContentWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "revision",
          "server_updated_at",
          "created"
        ],
        "title": "ProgressAck"
      },
      "ProgressConflict": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "server_revision": {
            "type": "integer",
            "title": "Server Revision"
          },
          "server_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Server Updated At"
          },
          "server_snapshot": {
            "$ref": "#/components/schemas/ProgressSnapshot"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "server_revision",
          "server_updated_at",
          "server_snapshot"
        ],
        "title": "ProgressConflict",
        "description": "Ответ 409: на сервере снимок новее. Клиент решает — слить или перезаписать\n(`?force=true`), прогресс при этом не теряется ни с одной стороны."
      },
      "ProgressSnapshot": {
        "properties": {
          "schema_version": {
            "type": "string",
            "title": "Schema Version",
            "default": "2.1"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision",
            "description": "Счётчик версий на клиенте: растёт при каждой выгрузке"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Время последнего изменения на устройстве"
          },
          "day": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Day",
            "description": "Текущий период («День N»)"
          },
          "player": {
            "$ref": "#/components/schemas/PlayerState"
          },
          "pet": {
            "$ref": "#/components/schemas/PetState"
          },
          "wallet": {
            "$ref": "#/components/schemas/WalletState"
          },
          "goal": {
            "$ref": "#/components/schemas/GoalState"
          },
          "inventory": {
            "items": {
              "$ref": "#/components/schemas/InventoryEntry"
            },
            "type": "array",
            "title": "Inventory",
            "description": "Рюкзачок"
          },
          "lessons": {
            "$ref": "#/components/schemas/LessonsState"
          },
          "owned": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Owned",
            "description": "Купленное навсегда: вещи для дома, курсы, украшения коллекции"
          },
          "streak": {
            "$ref": "#/components/schemas/StreakState"
          },
          "last_bonus_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Bonus Date",
            "description": "Когда выдан бонус за вход"
          },
          "settings": {
            "$ref": "#/components/schemas/ProfileSettings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "updated_at",
          "day",
          "player",
          "pet",
          "wallet",
          "goal"
        ],
        "title": "ProgressSnapshot",
        "description": "Полный снимок локального профиля."
      },
...
      "StoredProgress": {
        "properties": {
          "profile_id": {
            "type": "string",
            "format": "uuid",
            "title": "Profile Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "server_updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Server Updated At"
          },
          "snapshot": {
            "$ref": "#/components/schemas/ProgressSnapshot"
          }
        },
        "type": "object",
        "required": [
          "profile_id",
          "revision",
          "server_updated_at",
          "snapshot"
        ],
        "title": "StoredProgress"
      },
...
      "StreakState": {
        "properties": {
          "days": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Days",
            "description": "Огонёк: дней подряд с действием",
            "default": 0
          },
          "last_action_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Action Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "StreakState"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WalletState": {
        "properties": {
          "balance": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Balance",
            "description": "Монетки в кошельке"
          },
          "savings": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Savings",
            "description": "Монетки в копилке"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "balance",
          "savings"
        ],
        "title": "WalletState",
        "description": "Отрицательный баланс запрещён ТЗ 2.5.6 — ограничение `ge=0` здесь и есть\nэто правило: снимок с минусом сервер не примет (422)."
      },
...
    }
  },
...
}